backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.6
openai>=1.57.0
umap-learn>=0.5.7
numpy>=2.0.0
scipy>=1.14.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
